requirements.txt
flask 
azure-search-documents 
pandas
openai
requests
python-dotenv
azure-functions
openpyxl
opencensus-ext-azure 
opencensus-ext-logging